Move most of login code to login module
