+++ src/login/provider.rs
@@ -1,4 +1,3 @@
-
 use actix_web::Error;
 use log::error;
 use serde::Deserialize;
