+++ src/login/routes.rs
@@ -6,21 +6,19 @@ use tera::Tera;
 use serde::Deserialize;
 
 #[cfg(feature = "github-login")]
-use crate::login::github::Github;
+use crate::login::providers::github::Github;
 #[cfg(feature = "gitlab-login")]
-use crate::login::gitlab::GitLab;
+use crate::login::providers::gitlab::GitLab;
 #[cfg(feature = "google-login")]
-use crate::login::google::Google;
+use crate::login::providers::google::Google;
 #[cfg(feature = "microsoft-login")]
-use crate::login::microsoft::Microsoft;
-#[cfg(feature = "third-party-login")]
+use crate::login::providers::microsoft::Microsoft;
 use crate::login::provider::ThirdPartyUserInfo;
-#[cfg(feature = "third-party-login")]
 use crate::login::provider::LoginProvider;
 use crate::storage::DbPool;
 use log::{info, error};
 
-use crate::models::user::{User, NewUser};
+use crate::login::models::{User, NewUser};
 
 use crate::env;
 
@@ -46,7 +44,6 @@ async fn home(
     
     let mut platforms = Vec::<HashMap::<&str, String>>::new();
 
-    #[cfg(feature = "third-party-login")]
     let host = req.connection_info().host().to_string();
     #[cfg(feature = "github-login")]
     if env::var(env::ENV_GITHUB_APP_CLIENT_ID).is_ok() && env::var(env::ENV_GITHUB_APP_CLIENT_SECRET).is_ok() {
@@ -233,7 +230,6 @@ async fn login_callback(
     }
 }
 
-#[cfg(feature = "third-party-login")]
 pub fn user_login_route_config(cfg: &mut web::ServiceConfig) {
     #[cfg(feature = "github-login")]
     cfg.service(login_github_callback);
