+++ src/routes/mod.rs
@@ -1,4 +1,2 @@
 pub mod user;
 pub mod config;
-#[cfg(feature = "third-party-login")]
-pub mod login;
